+++ libs/components/src/lib/data-grid/data-grid-row.spec.ts
@@ -138,7 +138,7 @@ describe('vwc-data-grid-row', () => {
 
 	describe('row-focused event', () => {
 		it('should fire the focused event when one of the cells is focused', async () => {
-			const spy = jest.fn();
+			const spy = vi.fn();
 			element.addEventListener('row-focused', spy);
 			element.dispatchEvent(new FocusEvent('cell-focused'));
 			expect(spy).toHaveBeenCalled();
@@ -267,15 +267,17 @@ describe('vwc-data-grid-row', () => {
 
 	describe('a11y', () => {
 		it('should pass html a11y test', async () => {
-			element = (await fixture(`
+			const element = (await fixture(`
 				<div role="grid">
 					<${COMPONENT_TAG}>
 						<div role="gridcell"></div>
 					</${COMPONENT_TAG}>
 				</div>
-			`)) as DataGridRow;
-			element.ariaSelected = 'true';
-			await elementUpdated(element);
+			`)) as HTMLDivElement;
+			const row = element.querySelector('vwc-data-grid-row') as DataGridRow;
+			row.ariaSelected = 'true';
+
+			await elementUpdated(row);
 
 			expect(await axe(element)).toHaveNoViolations();
 		});