+++ libs/components/src/shared/foundation/button/button.spec.ts
@@ -546,7 +546,7 @@ describe('Foundation Button', () => {
 
 			await connect();
 
-			form.requestSubmit = jest.fn();
+			form.requestSubmit = vi.fn();
 
 			element.click();
 
@@ -565,7 +565,7 @@ describe('Foundation Button', () => {
 
 			await connect();
 
-			form.requestSubmit = jest.fn();
+			form.requestSubmit = vi.fn();
 
 			element.click();
 
@@ -651,7 +651,7 @@ describe('Foundation Button', () => {
 	});
 
 	describe("of 'disabled'", () => {
-		xit('should not propagate when clicked', async () => {
+		it.skip('should not propagate when clicked', async () => {
 			const { connect, disconnect, element, parent } = await setup();
 
 			element.disabled = true;